Add files via upload
A vibrating system with three degrees of freedom, M,K,C are known, and the load f3(t)=F*e**jwt is applied to the third degree of freedom only. According to the theory of modal analysis of multi-degree-of-freedom structures: 1. Draw H22(w), H12(w) 2. Find the structure's intrinsic frequencies and damping ratios by the frequency response curves